The panel will be reset on exit, all fields will be cleared
The JTaxtFields and JRadioButtons will be cleared on exit, and some specific JButtons will be disabled as intended.

ps: the 'CIN' and 'Apoge' combination is unique, meaning if the user enters a CIN with another Apoge an error will appear (that primary key must be unique).
since the Apoge is a 9 digits number, its unlikely to have it twice anyways.
---------------------
TODO:
--> might wanna add some feedback to the button when clicked