Certificate template is ready
I've created an HTML page with some CSS as a template for the Certificate document, with specific places to fill in the CIN and N°Apoge from the Database.
I also hooked up a library that converts HTML files to PDF files.

So basically, whats left is to make a template for the Transcript document as well.